refactor: do not render any components in App
