+++ src/components/Page/Layout.tsx
@@ -6,27 +6,14 @@
  *
  */
 
-import { CSSProperties, useMemo } from "react";
 import { Outlet } from "react-router-dom";
-import { useComponentSize } from "@ndla/hooks";
-import { styled } from "@ndla/styled-system/jsx";
 import { Footer } from "../Footer";
-
-const PageLayout = styled("div", {
-  base: {
-    display: "flex",
-    flexDirection: "column",
-    // The minimum page height should be 100vh - masthead height
-    minHeight: "calc(100vh - (var(--masthead-height, 72px)))",
-  },
-});
+import { PageLayout } from "../Layout/PageLayout";
 
 export const Layout = () => {
-  const { height } = useComponentSize("masthead");
-  const mastheadHeightVar = useMemo(() => ({ "--masthead-height": `${height}px` }) as CSSProperties, [height]);
   return (
     <>
-      <PageLayout style={mastheadHeightVar}>
+      <PageLayout>
         <Outlet />
       </PageLayout>
       <Footer />
