+++ src/containers/App/App.tsx
@@ -22,10 +22,7 @@ import "@fontsource/source-serif-pro/700.css";
 import "../../style/index.css";
 
 import { ReactElement } from "react";
-import { Helmet } from "react-helmet-async";
-import { useTranslation } from "react-i18next";
 import { Route, Routes } from "react-router-dom";
-import { PageContainer } from "@ndla/ui";
 import MediaPage from "./MediaPage";
 import SearchPage from "./SearchPage";
 import SubjectMatterPage from "./SubjectMatterPage";
@@ -43,7 +40,6 @@ import GlossPage from "../GlossPage/GlossPage";
 import H5PPage from "../H5PPage/H5PPage";
 import Login from "../Login/Login";
 import Logout from "../Logout/Logout";
-import Messages from "../Messages/Messages";
 import { MessagesProvider, useMessages } from "../Messages/MessagesProvider";
 import NdlaFilm from "../NdlaFilm/NdlaFilm";
 import NodeDiffPage from "../NodeDiff/NodeDiffPage";
@@ -57,48 +53,41 @@ import TaxonomyVersionsPage from "../TaxonomyVersions/TaxonomyVersionsPage";
 import WelcomePage from "../WelcomePage/WelcomePage";
 
 const App = () => {
-  const { t } = useTranslation();
-
   return (
     <ErrorBoundary>
       <MessagesProvider>
         <SessionProvider initialValue={getSessionStateFromLocalStorage()}>
           <AuthInitializer>
-            <PageContainer>
-              <Helmet meta={[{ name: "description", content: t("meta.description") }]} />
-              <Routes>
-                <Route element={<MastheadLayout />}>
-                  <Route path="/" element={<Layout />}>
-                    <Route index element={<WelcomePage />} />
-                    <Route path="login/*" element={<Login />} />
-                    <Route path="logout/*" element={<Logout />} />
-                    <Route path="/subjectpage/*" element={<PrivateRoute component={<Subjectpage />} />} />
-                    <Route path="search/*" element={<PrivateRoute component={<SearchPage />} />} />
-                    <Route path="subject-matter/*" element={<PrivateRoute component={<SubjectMatterPage />} />} />
-                    <Route
-                      path="/edit-markup/:draftId/:language/*"
-                      element={<PrivateRoute component={<EditMarkupPage />} />}
-                    />
-                    <Route path="/concept/*" element={<PrivateRoute component={<ConceptPage />} />} />
-                    <Route path="/gloss/*" element={<PrivateRoute component={<GlossPage />} />} />
-                    <Route path="/preview/:draftId/:language/*" element={<PreviewDraftPage />} />
-                    <Route path="/compare/:draftId/:language/*" element={<ComparePage />} />
-                    <Route path="/media/*" element={<PrivateRoute component={<MediaPage />} />} />
-                    <Route path="/film/*" element={<PrivateRoute component={<NdlaFilm />} />} />
-                    <Route path="/structure/*" element={<PrivateRoute component={<StructurePage />} />} />
-                    <Route path="/programme/*" element={<PrivateRoute component={<ProgrammePage />} />} />
-                    <Route path="/taxonomyVersions/*" element={<PrivateRoute component={<TaxonomyVersionsPage />} />} />
-                    <Route path="/nodeDiff/:nodeId" element={<PrivateRoute component={<NodeDiffPage />} />} />
-                    <Route path="/frontpage/" element={<PrivateRoute component={<FrontpageEditPage />} />} />
-                    <Route path="/forbidden" element={<ForbiddenPage />} />
-                    <Route path="*" element={<NotFoundPage />} />
-                  </Route>
-                  <Route path="/h5p/*" element={<PrivateRoute component={<H5PPage />} />} />
+            <Routes>
+              <Route element={<MastheadLayout />}>
+                <Route path="/" element={<Layout />}>
+                  <Route index element={<WelcomePage />} />
+                  <Route path="login/*" element={<Login />} />
+                  <Route path="logout/*" element={<Logout />} />
+                  <Route path="/subjectpage/*" element={<PrivateRoute component={<Subjectpage />} />} />
+                  <Route path="search/*" element={<PrivateRoute component={<SearchPage />} />} />
+                  <Route path="subject-matter/*" element={<PrivateRoute component={<SubjectMatterPage />} />} />
+                  <Route
+                    path="/edit-markup/:draftId/:language/*"
+                    element={<PrivateRoute component={<EditMarkupPage />} />}
+                  />
+                  <Route path="/concept/*" element={<PrivateRoute component={<ConceptPage />} />} />
+                  <Route path="/gloss/*" element={<PrivateRoute component={<GlossPage />} />} />
+                  <Route path="/preview/:draftId/:language/*" element={<PreviewDraftPage />} />
+                  <Route path="/compare/:draftId/:language/*" element={<ComparePage />} />
+                  <Route path="/media/*" element={<PrivateRoute component={<MediaPage />} />} />
+                  <Route path="/film/*" element={<PrivateRoute component={<NdlaFilm />} />} />
+                  <Route path="/structure/*" element={<PrivateRoute component={<StructurePage />} />} />
+                  <Route path="/programme/*" element={<PrivateRoute component={<ProgrammePage />} />} />
+                  <Route path="/taxonomyVersions/*" element={<PrivateRoute component={<TaxonomyVersionsPage />} />} />
+                  <Route path="/nodeDiff/:nodeId" element={<PrivateRoute component={<NodeDiffPage />} />} />
+                  <Route path="/frontpage/" element={<PrivateRoute component={<FrontpageEditPage />} />} />
+                  <Route path="/forbidden" element={<ForbiddenPage />} />
+                  <Route path="*" element={<NotFoundPage />} />
                 </Route>
-              </Routes>
-
-              <Messages />
-            </PageContainer>
+                <Route path="/h5p/*" element={<PrivateRoute component={<H5PPage />} />} />
+              </Route>
+            </Routes>
           </AuthInitializer>
         </SessionProvider>
       </MessagesProvider>
